Simplify and clean up produce implementation
The implementation was not producing the correct results,
as it was only doing the double hashing on the first level.
There seem to be some confusion in the code in the difference
between bytes, chars, and object.
The new implementation stores the sha256 object in an array,
and only call digest when necessary.